requirements.txt
matplotlib
mysql_connector_repackaged
numpy
pandas
scikit_learn
seaborn
plotly
pickle
sklearn
PIL